requirements.txt
uvicorn
fastapi[all]
pydantic
openai
langchain
langchain_openai
python-dotenv